Restructure core file tree for added organization
